+++ internal/repos/http.go
@@ -178,7 +178,7 @@ func (h *HttpRepo) ListCompletions(ctx context.Context, kind string, args []stri
 	switch kind {
 	case CompletionKindNames:
 		namePrefix, seg := longestPath(toComplete)
-		sr, err := h.List(ctx, model.ToFilters(nil, nil, nil, nil, &namePrefix,
+		sr, err := h.List(ctx, model.ToFilters(nil, nil, nil, nil, &namePrefix, nil,
 			&model.FilterOptions{NameFilterType: model.PrefixMatch}))
 		if err != nil {
 			return nil, err
@@ -206,7 +206,7 @@ func (h *HttpRepo) ListCompletions(ctx context.Context, kind string, args []stri
 		return vs, nil
 	case CompletionKindNamesOrIds:
 		namePrefix, seg := longestPath(toComplete)
-		sr, err := h.List(ctx, model.ToFilters(nil, nil, nil, nil, &namePrefix,
+		sr, err := h.List(ctx, model.ToFilters(nil, nil, nil, nil, &namePrefix, nil,
 			&model.FilterOptions{NameFilterType: model.PrefixMatch}))
 		if err != nil {
 			return nil, err